+++ killstats/api/killboard/corporation/killboard.py
@@ -7,15 +7,12 @@ from ninja import NinjaAPI
 
 # AA Killstats
 from killstats.api import schema
-from killstats.api.account_manager import AccountManager
-from killstats.api.helpers import KillboardDate, get_corporations
-from killstats.api.killboard_manager import (
-    killboard_dashboard,
-    killboard_hall,
-    killboard_process_kills,
+from killstats.api.killboard.killboard_helper import (
+    get_killmails_data,
+    get_killstats_halls,
+    get_killstats_stats,
 )
 from killstats.hooks import get_extension_logger
-from killstats.models.killboard import Killmail
 from killstats.models.killstatsaudit import CorporationsAudit
 
 logger = get_extension_logger(__name__)
@@ -26,59 +23,41 @@ class KillboardCorporationApiEndpoints:
 
     # pylint: disable=too-many-locals
     def __init__(self, api: NinjaAPI):
+
+        # Killmails
         @api.get(
-            "killboard/month/{month}/year/{year}/corporation/{corporation_id}/",
-            response={200: List[schema.KillboardIndex], 403: str},
+            "killmail/month/{month}/year/{year}/corporation/{corporation_id}/{mode}/",
+            response={200: dict, 403: str},
             tags=self.tags,
         )
-        def get_killstats(request, month, year, corporation_id: int):
-            # Killboard
-            if corporation_id == 0:
-                corporations = get_corporations(request)
-            else:
-                corporations = [corporation_id]
-
-            killmail_year = (
-                Killmail.objects.prefetch_related("victim", "victim_ship")
-                .filter(killmail_date__year=year)
-                .order_by("-killmail_date")
+        def get_corporation_killmails(
+            request, month, year, corporation_id: int, mode, page_size: int = 100
+        ):
+            return get_killmails_data(
+                request, month, year, corporation_id, mode, page_size, "corporation"
             )
 
-            killmail_filtered = killmail_year.filter_entities(corporations)
-
-            killmail_month = killmail_filtered.filter(
-                killmail_date__year=year,
-                killmail_date__month=month,
+        # Hall of Fame/Shame
+        @api.get(
+            "halls/month/{month}/year/{year}/corporation/{corporation_id}/",
+            response={200: List[schema.KillboardHall], 403: str},
+            tags=self.tags,
+        )
+        def get_corporation_halls(request, month, year, corporation_id: int):
+            return get_killstats_halls(
+                request, month, year, corporation_id, "corporation"
             )
 
-            kills, totalvalue, losses, totalvalue_loss = killboard_process_kills(
-                killmail_month, corporations
+        # Stats
+        @api.get(
+            "stats/month/{month}/year/{year}/corporation/{corporation_id}/",
+            response={200: List[schema.KillboardStats], 403: str},
+            tags=self.tags,
+        )
+        def get_corporation_killstats(request, month, year, corporation_id: int):
+            return get_killstats_stats(
+                request, month, year, corporation_id, "corporation"
             )
-
-            date = KillboardDate(month, year)
-
-            account = AccountManager(corporations=corporations)
-            mains, _ = account.get_mains_alts()
-
-            stats = killboard_dashboard(killmail_filtered, date, corporations)
-
-            shame, fame = killboard_hall(killmail_month, corporations, mains)
-
-            output = []
-            output.append(
-                {
-                    "kills": kills,
-                    "losses": losses,
-                    "totalKills": totalvalue,
-                    "totalLoss": totalvalue_loss,
-                    # Addons
-                    "shame": shame,
-                    "fame": fame,
-                    "stats": stats,
-                }
-            )
-
-            return output
 
         @api.get(
             "killboard/corporation/admin/",