+++ killstats/api/schema.py
@@ -16,13 +16,12 @@ class Killboard(Schema):
     attackers: Optional[list] = None
 
 
-class KillboardIndex(Schema):
-    kills: Optional[list] = None
-    losses: Optional[list] = None
-    totalKills: Optional[int] = None
-    totalLoss: Optional[int] = None
+class KillboardHall(Schema):
     shame: Optional[list] = None
     fame: Optional[list] = None
+
+
+class KillboardStats(Schema):
     stats: Optional[list] = None