+++ killstats/urls.py
@@ -3,28 +3,25 @@
 # Django
 from django.urls import path, re_path
 
+from killstats import views
+
 # AA Killstats
 from killstats.api import api
-from killstats.views import (
-    add_alliance,
-    add_corp,
-    alliance_admin,
-    corporation_admin,
-    killboard_index,
-)
 
 app_name: str = "killstats"
 
 urlpatterns = [
     # Killstats
-    path("index", killboard_index, {"corporation_pk": 0}, name="index"),
-    path("corporation/<int:corporation_pk>/", killboard_index, name="corporation"),
-    path("alliance/<int:alliance_pk>/", killboard_index, name="alliance"),
-    path("corporation_admin/", corporation_admin, name="corporation_admin"),
-    path("alliance_admin/", alliance_admin, name="alliance_admin"),
+    path("index", views.killboard_index, {"corporation_pk": 0}, name="index"),
+    path(
+        "corporation/<int:corporation_pk>/", views.killboard_index, name="corporation"
+    ),
+    path("alliance/<int:alliance_pk>/", views.killboard_index, name="alliance"),
+    path("corporation_admin/", views.corporation_admin, name="corporation_admin"),
+    path("alliance_admin/", views.alliance_admin, name="alliance_admin"),
     # -- Killstats Audit
-    path("add_corp/", add_corp, name="add_corp"),
-    path("add_alliance/", add_alliance, name="add_alliance"),
+    path("add_corp/", views.add_corp, name="add_corp"),
+    path("add_alliance/", views.add_alliance, name="add_alliance"),
     # -- API System
     re_path(r"^api/", api.urls),
 ]